[IMP] website, website_sale: improve header options
Miscellaneous improvements of the header options:
- Text updates (labels, tooltips, ...)
- Reordering of options
- Creation of sections to improve readability
- Separation of "mobile alignment" and "desktop alignment"

See also : odoo/design-themes#1128

task-4556047
