+++ addons/html_builder/static/src/sidebar/option_container.scss
@@ -48,6 +48,9 @@
     }
 }
 
-.o_we_img_animate:hover img {
-    content: image-set(var(--animate-src));
+.o_we_img_animate {
+    padding-top: 10px;
+    &:hover img {
+        content: image-set(var(--animate-src));
+    }
 }
