+++ addons/website/static/src/builder/plugins/options/header/header_box_option.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<templates xml:space="preserve">
+
+<t t-name="website.HeaderBoxOption">
+    <BorderConfigurator 
+        label.translate="Border" 
+        action="'styleActionHeader'" 
+        withRoundCorner="this.domState.withRoundCorner"
+    />
+    <ShadowOption 
+        setShadowModeAction="'setShadowModeHeader'" 
+        setShadowAction="'setShadowHeader'"
+    />
+</t>
+
+</templates>
