+++ addons/website/static/src/builder/plugins/options/header/header_font_option.js
@@ -1,0 +1,6 @@
+import { BaseOptionComponent } from "@html_builder/core/utils";
+
+export class HeaderFontOption extends BaseOptionComponent {
+    static template = "website.HeaderFontOption";
+    static props = {};
+}
