+++ addons/website/static/src/builder/plugins/options/header/header_icon_background_option.js
@@ -1,0 +1,6 @@
+import { BaseOptionComponent } from "@html_builder/core/utils";
+
+export class HeaderIconBackgroundOption extends BaseOptionComponent {
+    static template = "website.HeaderIconBackgroundOption";
+    static props = {};
+}
