+++ addons/website/static/src/builder/plugins/options/header/header_template_option.js
@@ -1,0 +1,6 @@
+import { BaseOptionComponent } from "@html_builder/core/utils";
+
+export class HeaderTemplateOption extends BaseOptionComponent {
+    static template = "website.HeaderTemplateOption";
+    static props = {};
+}
