--- addons/website/static/src/builder/plugins/options/header_border_option.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates xml:space="preserve">
-
-<t t-name="website.HeaderBorderOption">
-    <BuilderContext>
-        <BorderConfigurator label.translate="Border" action="'styleActionHeader'" withRoundCorner="this.domState.withRoundCorner"/>
-        <ShadowOption setShadowModeAction="'setShadowModeHeader'" setShadowAction="'setShadowHeader'"/>
-    </BuilderContext>
-</t>
-
-</templates>
