+++ addons/website/static/src/builder/plugins/options/website_page_config_option.xml
@@ -2,7 +2,7 @@
 <templates xml:space="preserve">
 
 <t t-name="website.TopMenuVisibilityOption">
-    <BuilderRow label.translate="Header Position" t-if="!this.isActiveItem('header_sidebar_opt')">
+    <BuilderRow label.translate="Header Position" level="1" t-if="!this.isActiveItem('header_sidebar_opt')">
         <BuilderSelect preview="false" action="'setWebsiteHeaderVisibility'">
             <BuilderSelectItem actionValue="'overTheContent'" id="'overTheContent'" t-if="props.doesPageOptionExist('header_overlay')">Over The Content</BuilderSelectItem>
             <BuilderSelectItem actionValue="'regular'">Regular</BuilderSelectItem>
@@ -10,10 +10,10 @@
         </BuilderSelect>
     </BuilderRow>    
     <t t-if="isActiveItem('overTheContent')">
-        <BuilderRow label.translate="Background" level="1" t-if="props.doesPageOptionExist('header_color')">
+        <BuilderRow label.translate="Background" level="2" t-if="props.doesPageOptionExist('header_color')">
             <BuilderColorPicker action="'setPageWebsiteDirty'" styleAction="'background-color'" enabledTabs="['custom']"/>
         </BuilderRow>
-        <BuilderRow label.translate="Text Color" level="1" t-if="props.doesPageOptionExist('header_text_color')">
+        <BuilderRow label.translate="Text Color" level="2" t-if="props.doesPageOptionExist('header_text_color')">
             <BuilderColorPicker action="'setPageWebsiteDirty'" styleAction="'color'" enabledTabs="['solid', 'custom']"/>
         </BuilderRow>
     </t>
