+++ addons/website_sale/static/src/website_builder/website_sale_show_empty_option_plugin.js
@@ -1,13 +1,14 @@
-import { HEADER_END } from "@website/builder/plugins/options/header_option_plugin";
+import { HEADER_ELEMENTS } from "@website/builder/plugins/options/header/header_option_plugin";
 import { Plugin } from "@html_editor/plugin";
 import { withSequence } from "@html_editor/utils/resource";
 import { registry } from "@web/core/registry";
+import { after } from "@html_builder/utils/option_sequence";
 
 class WebsiteSaleShowEmptyOptionPlugin extends Plugin {
     static id = "showEmptyOption";
     resources = {
         builder_options: [
-            withSequence(HEADER_END, {
+            withSequence(after(HEADER_ELEMENTS), {
                 template: "website_sale.ShowEmptyOption",
                 selector: "#wrapwrap > header",
                 editableOnly: false,
