+++ addons/website_sale_wishlist/static/src/website_builder/show_empty_option.xml
@@ -3,7 +3,7 @@
 
 <t t-name="website_sale_wishlist.ShowEmptyOption" t-inherit="website_sale.ShowEmptyOption" t-inherit-mode="extension">
     <BuilderRow id="'show_empty_opt'" position="inside">
-        <BuilderButton title.translate="Show/hide empty wishlist" icon="'fa-heart'"
+        <BuilderButton title.translate="Wishlist" icon="'fa-heart'"
             className="'flex-grow-1'"
             actionParam="{views: ['!website_sale_wishlist.header_hide_empty_wishlist_link']}"
         />
